cursor: send wlroots events for pressed up/down even when grabbed
In those cases there is no cursor focus, but wlroots can update its
internal state. This is needed for proper DnD operation.

The current implementation has a "bug": it sends button released events
in some cases where the button pressed wasn't sent. This results in a
wlroots error. This will be resolved in the future as wlroots will track
pressed buttons, and we'll be able to skip sending those released
events.